split code, add rule levels, updated workflows
