+++ src/rules/assessment/common/learning-objectives.js
@@ -6,5 +6,5 @@ export default {
     return !learningObjectives || learningObjectives.indexOf('SWBAT') !== 0 ?
       '"Learning Objectives" field should not be blank, and it should start with “SWBAT”.' : undefined
   },
-  level: RULE_LEVELS.ISSUE
+  level: RULE_LEVELS.SUGGESTION
 }
