+++ src/rules/assessment/multiple-choice-questions/shuffle-answers.js
@@ -1,8 +1,8 @@
-import {ASSESSMENT_TYPES, RULE_LEVELS} from '../../const'
+import {ASSESSMENT, RULE_LEVELS} from '../../const'
 
 export default {
   action: (assessment) => {
-    return assessment.type === ASSESSMENT_TYPES.MULTIPLE_CHOICE && !assessment.source.isRandomized ?
+    return assessment.type === ASSESSMENT.TYPES.MULTIPLE_CHOICE && !assessment.source.isRandomized ?
       '"Shuffle Answers" in "Multiple Choice Questions" should be toggled on' : undefined
   },
   level: RULE_LEVELS.ISSUE
