+++ src/rules/assessment/parsons-puzzle/code-to-become-blocks.js
@@ -1,8 +1,8 @@
-import {ASSESSMENT_TYPES, RULE_LEVELS} from '../../const'
+import {ASSESSMENT, RULE_LEVELS} from '../../const'
 
 export default {
   action: (assessment) => {
-    return assessment.type === ASSESSMENT_TYPES.PARSONS_PUZZLE && !assessment.source.initial ?
+    return assessment.type === ASSESSMENT.TYPES.PARSONS_PUZZLE && !assessment.source.initial ?
       '"Code to Become Blocks" in "Parsons Puzzle" should not be blank.' : undefined
   },
   level: RULE_LEVELS.ISSUE
