+++ src/rules/const.js
@@ -1,21 +1,10 @@
+import {TYPES} from './assessment/const'
 export const RULE_LEVELS = {
   SUGGESTION: 0,
   WARNING: 1,
   ISSUE: 2
 }
 
-export const ASSESSMENT_TYPES = {
-  TEST: 'test',
-  MULTIPLE_CHOICE: 'multiple-choice',
-  FILL_IN_THE_BLANKS: 'fill-in-the-blanks',
-  FREE_TEXT: 'free-text',
-  FREE_TEXT_AUTO: 'free-text-auto',
-  CODE_OUTPUT_COMPARE: 'code-output-compare',
-  CUSTOM: 'custom',
-  MATH: 'math-stack',
-  GRADE_BOOK: 'grade-book',
-  PARSONS_PUZZLE: 'parsons-puzzle',
-  JUPYTER_NOTEBOOK: 'jupyter-notebook',
-  SENSE_NETWORK: 'sense-network',
-  RANDOM: 'random'
+export const ASSESSMENT = {
+  TYPES
 }
