+++ src/ui/icons.js
@@ -1,0 +1,17 @@
+import {RULE_LEVELS} from '../rules/const'
+
+export const getIconByLevel = (level) => {
+  switch (level) {
+    case RULE_LEVELS.SUGGESTION:
+      /* eslint-disable-next-line max-len */
+      return '<svg xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 128 128"><g fill="#40c0e7"><path d="M68.97 31.66c-.64-.49-1.38-.85-2.24-1.05c-.85-.19-1.75-.3-2.72-.3c-.97 0-1.88.1-2.74.3c-.87.19-1.61.55-2.26 1.05c-.64.49-1.14 1.14-1.51 1.95c-.36.81-.55 1.83-.55 3.04c0 1.19.19 2.19.55 3.02c.37.81.87 1.48 1.51 1.97s1.39.85 2.26 1.06c.86.21 1.77.32 2.74.32c.96 0 1.87-.11 2.72-.32c.86-.21 1.6-.57 2.24-1.06c.64-.49 1.15-1.16 1.53-1.97c.38-.83.57-1.83.57-3.02c0-1.22-.19-2.24-.57-3.04c-.38-.81-.89-1.46-1.53-1.95zM57.52 48.94h12.97v47.51H57.52z"/><path d="M64 .41C28.93.41.41 28.93.41 64c0 35.06 28.52 63.59 63.58 63.59S127.58 99.07 127.58 64C127.59 28.93 99.06.41 64 .41zm0 118C34 118.41 9.6 94 9.6 64C9.6 34 34 9.59 64 9.59S118.41 34 118.41 64.01C118.4 94 93.99 118.41 64 118.41z"/></g></svg>'
+    case RULE_LEVELS.WARNING:
+      /* eslint-disable-next-line max-len */
+      return '<svg xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 128 128"><path fill="#fcc21b" d="M9.6 114.26L64 20.03l54.41 94.23z"/><g fill="#2f2f2f"><path d="M127.07 115.93L66.89 11.7c-.6-1.03-1.7-1.67-2.89-1.67c-1.19 0-2.29.64-2.88 1.67L.93 115.93c-.6 1.03-.6 2.3 0 3.34c.6 1.03 1.7 1.66 2.89 1.66h120.36c1.19 0 2.29-.63 2.89-1.66c.59-1.04.59-2.31 0-3.34zM9.6 114.26L64 20.03l54.41 94.23H9.6z"/><path d="M59.12 83.21h9.76c.36 0 .65-.23.67-.53l2.66-38.27c.01-.16-.05-.3-.18-.42a.736.736 0 0 0-.49-.18H56.46c-.19 0-.37.07-.49.18a.54.54 0 0 0-.18.42l2.67 38.27c.01.29.3.53.66.53zm-3.01 20.11c0-1.41.2-2.6.61-3.57c.4-.98.96-1.77 1.68-2.37c.72-.6 1.56-1.04 2.54-1.31c.97-.27 2.01-.41 3.13-.41c1.05 0 2.05.14 3 .41c.95.27 1.78.71 2.49 1.31c.72.6 1.29 1.38 1.71 2.37c.42.97.63 2.16.63 3.57c0 1.35-.21 2.51-.63 3.46c-.42.96-.99 1.76-1.71 2.39c-.71.63-1.54 1.09-2.49 1.37c-.95.28-1.95.43-3 .43c-1.11 0-2.15-.14-3.13-.43s-1.82-.74-2.54-1.37a6.196 6.196 0 0 1-1.68-2.39c-.41-.96-.61-2.11-.61-3.46z"/></g></svg>'
+    case RULE_LEVELS.ISSUE:
+      /* eslint-disable-next-line max-len */
+      return '<svg xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 128 128"><path fill="#ed6c30" d="m79.64 64l29.9-29.9c1.38-1.38 1.38-3.64 0-5.03L98.93 18.46a3.573 3.573 0 0 0-5.03 0L64 48.36l-29.9-29.9a3.573 3.573 0 0 0-5.03 0l-10.6 10.61a3.555 3.555 0 0 0 0 5.03L48.36 64l-29.9 29.9c-1.38 1.38-1.38 3.64 0 5.03l10.61 10.61a3.573 3.573 0 0 0 5.03 0L64 79.63l29.9 29.91a3.573 3.573 0 0 0 5.03 0l10.61-10.61c1.38-1.39 1.38-3.65 0-5.03L79.64 64z"/></svg>'
+    default:
+      return ''
+  }
+}
